+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 ## Added
 
-## Fixed
-
-## Changed
+- Added combined "path-handler" for a more convenient and encapsulated API
 
 # 0.0.25 (2021-03-28 / ce0c699)